requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --resolver=backtracking -o requirements.txt
annotated-types==0.7.0
    # via pydantic
attrs==25.3.0
    # via
    #   frictionless
    #   jsonlines
    #   jsonschema
    #   referencing
beautifulsoup4==4.13.4
    # via bs4
bs4==0.0.2
    # via hdx-scraper-views (pyproject.toml)
certifi==2025.7.14
    # via requests
chardet==5.2.0
    # via frictionless
charset-normalizer==3.4.2
    # via requests
ckanapi==4.8
    # via hdx-python-api
click==8.2.1
    # via typer
defopt==7.0.0
    # via hdx-python-api
dnspython==2.7.0
    # via email-validator
docopt==0.6.2
    # via
    #   ckanapi
    #   num2words
docutils==0.21.2
    # via defopt
email-validator==2.2.0
    # via hdx-python-api
et-xmlfile==2.0.0
    # via openpyxl
frictionless==5.18.1
    # via hdx-python-utilities
hdx-python-api==6.4.6
    # via hdx-scraper-views (pyproject.toml)
hdx-python-country==3.9.6
    # via hdx-python-api
hdx-python-utilities==3.8.8
    # via
    #   hdx-scraper-views (pyproject.toml)
    #   hdx-python-api
    #   hdx-python-country
humanize==4.12.3
    # via frictionless
idna==3.10
    # via
    #   email-validator
    #   requests
ijson==3.4.0
    # via hdx-python-utilities
inflect==7.5.0
    # via quantulum3
isodate==0.7.2
    # via frictionless
jinja2==3.1.6
    # via frictionless
jsonlines==4.0.0
    # via hdx-python-utilities
jsonpath-ng==1.7.0
    # via libhxl
jsonschema==4.25.0
    # via
    #   frictionless
    #   tableschema-to-template
jsonschema-specifications==2025.4.1
    # via jsonschema
libhxl==5.2.2
    # via
    #   hdx-python-api
    #   hdx-python-country
loguru==0.7.3
    # via hdx-python-utilities
makefun==1.16.0
    # via hdx-python-api
markdown-it-py==3.0.0
    # via rich
marko==2.1.4
    # via frictionless
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.7.0
    # via inflect
num2words==0.5.16
    # via quantulum3
openpyxl==3.1.5
    # via hdx-python-utilities
petl==1.7.17
    # via frictionless
ply==3.11
    # via
    #   jsonpath-ng
    #   libhxl
pockets==0.9.1
    # via sphinxcontrib-napoleon
pydantic==2.11.7
    # via frictionless
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via rich
pyphonetics==0.5.3
    # via hdx-python-utilities
python-dateutil==2.9.0.post0
    # via
    #   frictionless
    #   hdx-python-utilities
    #   libhxl
python-io-wrapper==0.3.1
    # via libhxl
python-slugify==8.0.4
    # via
    #   ckanapi
    #   frictionless
pyyaml==6.0.2
    # via
    #   frictionless
    #   tableschema-to-template
quantulum3==0.9.2
    # via hdx-python-api
ratelimit==2.2.1
    # via hdx-python-utilities
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   ckanapi
    #   frictionless
    #   hdx-python-api
    #   libhxl
    #   requests-file
requests-file==2.1.0
    # via hdx-python-utilities
rfc3986==2.0.0
    # via frictionless
rich==14.1.0
    # via typer
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.14
    # via hdx-python-utilities
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
setuptools==80.9.0
    # via ckanapi
shellingham==1.5.4
    # via typer
simpleeval==1.0.3
    # via frictionless
simplejson==3.20.1
    # via ckanapi
six==1.17.0
    # via
    #   ckanapi
    #   pockets
    #   python-dateutil
    #   sphinxcontrib-napoleon
soupsieve==2.7
    # via beautifulsoup4
sphinxcontrib-napoleon==0.7
    # via defopt
structlog==25.4.0
    # via libhxl
tableschema-to-template==0.0.13
    # via hdx-python-utilities
tabulate==0.9.0
    # via frictionless
tenacity==9.1.2
    # via hdx-python-country
text-unidecode==1.3
    # via python-slugify
typeguard==4.4.4
    # via inflect
typer==0.16.0
    # via frictionless
typing-extensions==4.14.1
    # via
    #   beautifulsoup4
    #   frictionless
    #   pydantic
    #   pydantic-core
    #   referencing
    #   typeguard
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
unidecode==1.4.0
    # via
    #   libhxl
    #   pyphonetics
urllib3==2.5.0
    # via
    #   libhxl
    #   requests
validators==0.35.0
    # via frictionless
wheel==0.45.1
    # via libhxl
xlrd==2.0.2
    # via hdx-python-utilities
xlrd3==1.1.0
    # via libhxl
xlsx2csv==0.8.4
    # via hdx-python-utilities
xlsxwriter==3.2.5
    # via tableschema-to-template
xlwt==1.3.0
    # via hdx-python-utilities
